Add a to_hash on PullRequest to facilitate use in templates
Unfortunately, the method_missing magic we're doing in Base doesn't
seem to play nicely with things like Mustache, so we provide a to_hash
that can be used to get the data into a suitable format for them.

We'll probably need to move this into Base eventually (or figure out
something better than the method_missing magic).