Be more strict about method arguments when listing repositories
By checking whether the OctocatHerder::Connection we'll be using will
be making authenticated requests or not, we can fail earlier, with a
clear error message, rather than relying on the API request to fail
with a permission error.

Also, we're now checking to make sure that the repository type is one
of the acceptable values, rather than relying on the API request to
fail.